Factor out ancestor class Field of Attribute and Reflection.
